fix: pkValues npe and merge BatchInsertExecutorTest2
